+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #
 - Add support for env variables in git source url
+- Fix builds using podman runtime
 
 # 0.9.0
 - Rename `--docker-uri` option to `--runtime-uri` and make it global (no longer part of build and init command)